@@ -1,5 +1,13 @@
+<script lang="ts">
+	interface Props {
+		children?: import('svelte').Snippet;
+	}
+
+	let { children }: Props = $props();
+</script>
+
 <div class="legbox">
-	<slot />
+	{@render children?.()}
 </div>
 
 <style lang="scss">
